refactor(ui): extract feed/meta.ts — feedMetaText summary-line builder
